+++ crates/protocol/src/input.rs
@@ -9,9 +9,9 @@ pub struct PlayerInput {
     // World-space direction the ship should face.
     pub aim: Vec2,
     // Cumulative clicks survive repeated/missing simulation ticks without automatic fire.
-    pub blaster_clicks: u32,
+    pub blaster_clicks: u8,
     // Cumulative R presses, consumed once even across repeated simulation ticks.
-    pub blaster_reload_requests: u32,
+    pub blaster_reload_requests: u8,
     // Held state; the visual beam disappears when Space is released.
     pub phase_beam: bool,
 }
